+++ HISTORY.rst
@@ -12,10 +12,11 @@
 * Add custom DatabaseWarning to tag the SQLite usage warning.
 * Add keyword arguments to add_to_class instances.
 * Move add_to_class function to their own module called methods.py
-* Remove catch all exception handling for the check in and 
+* Remove catch all exception handling for the check in and
   check out views.
 * Improve checkouts tests code reducing redundant code.
-* Change how the HOME_VIEW setting is defined. 
+* Change how the HOME_VIEW setting is defined.
+* Remove the role permission grant and revoke permission.
 
 3.1.11 (2019-04-XX)
 ===================